+++ LibreNMS/Authentication/MysqlAuthorizer.php
@@ -104,7 +104,7 @@ class MysqlAuthorizer extends AuthorizerBase
             $user_id = $new_user->user_id;
 
             // set auth_id
-            $new_user->auth_id = $user_id;
+            $new_user->auth_id = $this->getUserid($username);
             $new_user->save();
 
             if ($user_id) {
